backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.20.0
aiosqlite>=0.18.0
pytest>=7.0.0
httpx>=0.24.0
aiomysql>=0.2.0
PyMySQL>=1.1.0
pydantic
google-generativeai
python-dotenv
pytest
httpx
requests
streamlit
